Add environment option to project initialization and refactor GitHub API client methods
